Rename package, get test working
