+++ geoprocessing/core.py
@@ -3,8 +3,8 @@ import json
 import os
 import shutil
 from six import string_types
-import gaia.inputs
-import gaia.processes
+import geoprocessing.inputs
+import geoprocessing.processes
 
 base_dir = os.path.join(
     os.path.dirname(os.path.realpath(__file__)))
@@ -26,8 +26,8 @@ class GaiaRequestParser(object):
 
     def __init__(self, process_name, data=None, parse=True, config=None):
         if config:
-            gaia.core.CONFIG = config
-        self.process = gaia.processes.base.create_process(process_name)
+            geoprocessing.core.CONFIG = config
+        self.process = geoprocessing.processes.base.create_process(process_name)
         if data and parse:
             self.parse(data)
 
@@ -37,8 +37,8 @@ class GaiaRequestParser(object):
         inputs = data['data_inputs']
         self.process.inputs = []
         for input in inputs:
-            gi = gaia.inputs.base.GaiaInput(input)
-            gi.io = gaia.inputs.base.create_io(inputs[input])
+            gi = geoprocessing.inputs.base.GaiaInput(input)
+            gi.io = geoprocessing.inputs.base.create_io(inputs[input])
             self.process.inputs.append(gi)
         return self.process
 
@@ -49,8 +49,8 @@ def getConfig():
     such as database connections
     :return: configuration
     """
-    if gaia.core.CONFIG is None:
-        config_file = os.path.join(base_dir, 'conf/gaia.local.cfg')
+    if geoprocessing.core.CONFIG is None:
+        config_file = os.path.join(base_dir, 'conf/geoprocessing.local.cfg')
         if not os.path.exists(config_file):
             shutil.copy(config_file.replace('local', 'dist'), config_file)
         parser = ConfigParser.ConfigParser()
@@ -60,17 +60,5 @@ def getConfig():
             config_dict[section] = {}
             for key, val in parser.items(section):
                 config_dict[section][key] = val
-        gaia.core.CONFIG = config_dict
-    return gaia.core.CONFIG
-
-
-if __name__ == '__main__':
-    test_dir = '../plugin_tests/data/geoprocess'
-    with open(os.path.join(base_dir,
-              '{}/within_nested_buffer_process.json'.format(test_dir))) as inf:
-        json_string = inf.read().replace(
-            '{basepath}', os.path.abspath(os.path.join(base_dir, test_dir)))
-    jsondata = json.loads(json_string)
-    grp = GaiaRequestParser('within', jsondata)
-    grp.process.calculate()
-    print grp.process.output
+        geoprocessing.core.CONFIG = config_dict
+    return geoprocessing.core.CONFIG
