+++ plugin.cmake
@@ -1,3 +1,3 @@
-add_python_test(simplewps PLUGIN gaia)
+add_python_test(geoprocess PLUGIN gaia)
 add_python_style_test(pep8_style_gaia_rest
                       "${PROJECT_SOURCE_DIR}/plugins/gaia/server/rest")
